examples/siamese_network/requirements.txt
skeltorch~=1.2.0
torchaudio~=0.7.2
